refactor(charts): move x-zoom into a component folder
Private components in this package live in folders with an index
barrel and a colocated test directory (center, chart-layout, ...);
x-zoom was the odd one out as flat files with its test in a shared
test dir. Consumers keep importing ../private/x-zoom, which now
resolves to the folder index.
