+++ projects/js-packages/charts/src/charts/private/x-zoom/index.ts
@@ -1,0 +1,1 @@
+export { useXZoom, ZoomClip, ZoomResetButton, ZoomSelectionRect } from './x-zoom';
