+++ projects/js-packages/charts/src/charts/private/x-zoom/x-zoom.module.scss
